move user seeder
