+++ src/seeder.ts
@@ -1,6 +1,6 @@
 import { seeder } from 'nestjs-seeder';
 import { DatabaseModule } from './database/database.module';
-import { UsersSeeder } from './users/entities/user.seeder';
+import { UsersSeeder } from './users/user.seeder';
 import { UsersModule } from './users/users.module';
 import { RolesModule } from './roles/roles.module';
 import { ConfigModule } from './config/config.module';
